Rename piet-wgsl to vello!
Also drops the scene crate and moves the code to the main project.
