--- piet-scene/src/scene/resource.rs
@@ -1,31 +1,0 @@
-use core::ops::Range;
-use peniko::ColorStop;
-
-#[derive(Default)]
-/// Collection of late bound resources for a scene or scene fragment.
-pub struct ResourceBundle {
-    /// Sequence of resource patches.
-    pub patches: Vec<ResourcePatch>,
-    /// Cache of color stops, referenced by range from the patches.
-    pub stops: Vec<ColorStop>,
-}
-
-impl ResourceBundle {
-    /// Clears the resource set.
-    pub(crate) fn clear(&mut self) {
-        self.patches.clear();
-        self.stops.clear();
-    }
-}
-
-#[derive(Clone)]
-/// Description of a late bound resource.
-pub enum ResourcePatch {
-    /// Gradient ramp resource.
-    Ramp {
-        /// Byte offset to the ramp id in the draw data stream.
-        offset: usize,
-        /// Range of the gradient stops in the resource set.
-        stops: Range<usize>,
-    },
-}
