--- piet-wgsl/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "piet-wgsl"
-version = "0.1.0"
-authors = ["Raph Levien <raph@google.com>"]
-license = "MIT/Apache-2.0"
-edition = "2021"
-
-[dependencies]
-wgpu = "0.14"
-raw-window-handle = "0.5"
-futures-intrusive = "0.5.0"
-parking_lot = "0.12"
-bytemuck = { version = "1.12.1", features = ["derive"] }
-piet-scene = { path = "../piet-scene" }
