+++ vello/examples/winit/src/test_scene.rs
@@ -1,7 +1,8 @@
 use crate::pico_svg::PicoSvg;
 use crate::simple_text::SimpleText;
-use piet_scene::kurbo::{Affine, BezPath, Ellipse, PathEl, Point, Rect};
-use piet_scene::*;
+use vello::kurbo::{Affine, BezPath, Ellipse, PathEl, Point, Rect};
+use vello::peniko::*;
+use vello::*;
 
 pub fn render_funky_paths(sb: &mut SceneBuilder) {
     use PathEl::*;
