+++ vello/shader/bbox_clear.wgsl
