+++ vello/shader/clip_reduce.wgsl
