+++ vello/shader/path_coarse.wgsl
