+++ vello/shader/path_coarse_full.wgsl
