+++ vello/shader/shared/clip.wgsl
